+++ packages/common/CHANGELOG.md
@@ -8,7 +8,8 @@
 - `createCredentialV1`, to create credentials using a seedPhrase.
 - Added `sendCredentialDeployment` to send credentials created from `createCredentialV1` to the chain.
 - `getSignedCredentialDeploymentTransactionHash` to get the transaction hash of credentials created from `createCredentialV1`.
--  Added `ConfigureBaker` to `AccountTransactionType` enum.
+- Added `ConfigureBaker` to `AccountTransactionType` enum.
+- Added `ConcordiumHdWallet` with functions to get keys and randomness from a seedPhrase.
 
 ## 2.3.2 2022-7-26