+++ packages/rust-bindings/CHANGELOG.md
@@ -6,3 +6,4 @@
 
 - `createCredentialV1`
 - `createIdRequestV1`
+-  Bindings for the HdWallet methods: `getAccountSigningKey`, `getAccountPublicKey`, `getPrfKey`, `getSignatureBlindingRandomness` and `getAttributeCommitmentRandomness`.